# Conflicts:
#	x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/security/authz/RoleRestrictionTests.java
#	x-pack/plugin/security/qa/security-trial/src/javaRestTest/java/org/elasticsearch/xpack/security/apikey/ApiKeyRestIT.java

+++ server/src/main/java/org/elasticsearch/script/VectorScoreScriptUtils.java
@@ -99,7 +99,7 @@ public class VectorScoreScriptUtils {
 
             if (normalizeQuery) {
                 for (int dim = 0; dim < this.queryVector.length; dim++) {
-                    this.queryVector[dim] /= queryMagnitude;
+                    this.queryVector[dim] /= (float) queryMagnitude;
                 }
             }
         }
